split get*Filter into ez-query-helpers.ts from advanced-search.tsx for faster unit tests
