+++ app/javascript/advanced-search.tsx
@@ -4,9 +4,7 @@ import Cookie from "js-cookie"
 import DanceTable from "./dance-table"
 import EzCheckboxFilter from "./ez-checkbox-filter"
 import Filter from "./filter"
-
-const matchNothing: Filter = ["or"] //  kinda non-obvious how to express tihs in filters, eh?
-const matchEverything: Filter = ["and"] //  kinda non-obvious how to express tihs in filters, eh?
+import { getVerifiedFilter, getPublishFilter } from "./ez-query-helpers"
 
 export const AdvancedSearch = () => {
   const [signedIn] = useState(() => Cookie.get("signed_in"))
@@ -86,75 +84,6 @@ export const AdvancedSearch = () => {
       <DanceTable filter={grandFilter} />
     </div>
   )
-}
-
-export const getVerifiedFilter = ({
-  v,
-  nv,
-  vbm,
-  nvbm,
-}: {
-  v: boolean
-  nv: boolean
-  vbm: boolean
-  nvbm: boolean
-}): Filter => {
-  const r: Filter = ["or"]
-  if (v) {
-    if (nv) {
-      return matchEverything
-    } else {
-      r.push(["compare", ["constant", 0], "<", ["tag", "verified"]])
-    }
-  } else {
-    if (nv) {
-      nvbm || r.push(["compare", ["constant", 0], "=", ["tag", "verified"]])
-    } else {
-      // do nothing
-    }
-  }
-  if (vbm) {
-    if (nvbm) {
-      return matchEverything
-    } else {
-      v || r.push(["my tag", "verified"])
-    }
-  } else {
-    if (nvbm) {
-      r.push(["no", ["my tag", "verified"]])
-    } else {
-      // do nothing
-    }
-  }
-  return r
-}
-
-export const getPublishFilter = ({
-  all = false,
-  sketchbook = false,
-  off = false,
-}: {
-  all?: boolean
-  sketchbook?: boolean
-  off?: boolean
-}): Filter => {
-  const sbk = sketchbook
-  const allFilters: Filter[] = all ? [["publish", "all"]] : []
-  const offFilters: Filter[] = off ? [["publish", "off"]] : []
-  const sbkFilters: Filter[] = sbk ? [["publish", "sketchbook"]] : []
-  const filters: Filter[] = [...allFilters, ...sbkFilters, ...offFilters]
-  switch (filters.length) {
-    case 0:
-      return matchNothing
-    case 1:
-      return filters[0]
-    case 2:
-      return ["or", ...filters]
-    case 3:
-      return matchEverything
-    default:
-      throw new Error("fell through case statement")
-  }
 }
 
 export default AdvancedSearch
